+++ doc/examples/skeleton_behaviour.py
@@ -31,10 +31,15 @@ class Foo(py_trees.Behaviour):
 
         What to do here?
           Delayed one-time initialisation that would otherwise interfere
-          with offline rendering of this behaviour in a tree to dot graph.
+          with offline rendering of this behaviour in a tree to dot graph
+          or validation of the behaviour's configuration.
+          
           Good examples include:
+          
           - Hardware or driver initialisation
           - Middleware initialisation (e.g. ROS pubs/subs/services)
+          - A parallel checking for a valid policy configuration after
+            children have been added or removed
         """
         self.logger.debug("  %s [Foo::setup()]" % self.name)