+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Release Notes
 
 **Breaking API**
 
+* [behaviour] setup() no longer returns a boolean, catch exceptions instead, #143.
 * [decorators] new-style decorators found in py_trees.decorators
 * [decorators] new-style decorators stop their running child on completion (SUCCESS||FAILURE)
 * [decorators] old-style decorators in py_trees.meta deprecated